+++ eva/src/components/dashPanelBoard.vue
@@ -275,6 +275,7 @@
                   icon
                   v-on="on"
                   @click="item.onClick"
+                  @click.middle="item.onMiddleClick"
                 >
                   <v-icon
                     class="edit icon-aut"
@@ -1061,6 +1062,10 @@ export default {
           label: 'Редактировать',
           icon: mdiAccountEdit,
           onClick: this.edit,
+          onMiddleClick: () => {
+            const routeData = this.$router.resolve('/profile');
+            window.open(routeData.href, '_blank');
+          },
           hide: this.inside,
         },
         {